[REFACTOR] : ♻️ MentionedTab interface 통합
