+++ src/components/Tag/Tag.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react';
 import * as UI from './style';
 import Tags from './Tags/Tags';
 import { useMentionedSchedules } from '../../api/hooks/Tag/useGetTag';
-import { TagElement } from './interfaces';
+import { Mention } from './interfaces';
 import { TagBlockCssProps } from './interfaces';
 import Board from '../Board/Board';
 import CalendarIcon from '../../assets/Icons/CalendarIcon';
@@ -47,7 +47,7 @@ const Tag = ({ types }: TagBlockCssProps) => {
   return (
     <Board icon={icon} title="tag" types={types}>
       <UI.StInsideBlock ref={targetDiv}>
-        {tags.map((tag: TagElement) => {
+        {tags.map((tag: Mention) => {
           return <Tags key={tag.mentionId} tag={tag} types={types} />;
         })}
       </UI.StInsideBlock>
